feat: move parser and analyzer inside the compiler directory
