+++ src/compiler/parser/tests/sql_samples/delete.txt
