+++ src/compiler/parser/tests/sql_samples/insert.txt
