+++ src/schemas/SchemaGenerator.py
@@ -1,3 +1,9 @@
+import sys
+import os
+
+dir_path = os.path.dirname(os.path.realpath(__file__))
+sys.path.append(os.path.abspath(os.path.join(dir_path, os.pardir)))
+
 class SchemaGenerator :
   def __init__ (self, **kwargs):
     self.name = ""
